Backend/requirements.txt
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.2.post1
anytree==2.12.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.8
async-lru==2.0.4
attrs==25.1.0
autograd==1.6.2
autopep8==1.7.0
babel==2.16.0
backcall==0.2.0
beartype==0.17.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
community==1.0.0b1
cycler==0.11.0
data-science-types==0.2.23
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==0.10.0
fastapi==0.115.3
fastjsonschema==2.16.1
flake8==5.0.4
flake8-docstrings==1.6.0
flake8-json==21.7.0
Flask==2.2.2
fonttools==4.36.0
fqdn==1.5.1
future==0.18.3
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.30
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
graspologic==3.3.0
graspologic-native==1.2.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.31.0
httpx==0.28.1
hyppo==0.4.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.1
jupyterlab_server==2.27.3
kiwisolver==1.4.4
llvmlite==0.42.0
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
mypy==0.971
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.2.1
notebook==6.4.12
notebook_shim==0.2.4
numba==0.59.0
numpy==1.23.2
overrides==7.7.0
packaging==24.2
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==4.3.6
pluggy==1.0.0
POT==0.9.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.0
pydocstyle==6.1.1
pyflakes==2.5.0
pygame==2.6.0
Pygments==2.14.0
Pympler==1.0.1
pynndescent==0.5.11
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-reportlog==0.1.2
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.2.1
pytoolconfig==1.2.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.3.1
QtPy==2.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rope==1.3.0
rpds-py==0.22.3
rsa==4.9.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
semver==2.13.0
Send2Trash==1.8.3
shelljob==0.6.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.36
stack-data==0.4.0
starlette==0.41.0
statsmodels==0.14.1
streamlit==1.18.1
tabulate==0.8.10
terminado==0.15.0
threadpoolctl==3.2.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.66.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.15.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.5
uri-template==1.3.0
uritemplate==4.2.0
urllib3==1.26.14
uvicorn==0.32.0
validators==0.20.0
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.1
zipp==3.13.0
